Build Payment confirmation screen and save property screen

+++ lib/scr/model/property_model.dart
@@ -5,7 +5,7 @@ class PropertyModel {
   final List<dynamic> images;
   final String? propertyCategory;
   final String? propertyType;
-  final int pricePerSqFt;
+  final double pricePerSqFt;
 
   PropertyModel({
     required this.id,
@@ -22,7 +22,7 @@ class PropertyModel {
       id: json["id"],
       title: json["title"],
       images: json["images"],
-      pricePerSqFt: (json["PricePerSqFt"] as num).toInt(),
+      pricePerSqFt: (json["PricePerSqFt"] as num).toDouble(),
       description:
           json["description"] ??
           "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book.",
